# Optimize walk performance