--- a/src/utils.jl
+++ b/src/utils.jl
@@ -89,3 +89,42 @@ function optional(ex::Expr)
 
     return funcs
 end
+
+"""
+    walk_tz_dir(f, dir) -> Nothing
+
+Walks the directory tree of a directory containing time zone information
+(e.g. `/usr/share/zoneinfo`). For each file encountered the function `f` with be called with
+the arguments `name` and `path`.
+
+## Examples
+
+Determine names of time zones in `/usr/share/zoneinfo`:
+
+```julia
+tz_names = String[]
+walk_tz_dir("/usr/share/zoneinfo") do name, path
+    open(path) do io
+        read(io, 4) == b"TZif" && push!(tz_names, name)
+    end
+end
+```
+"""
+function walk_tz_dir(f, dir)
+    check = Tuple{String,String}[("", dir)]
+    while !isempty(check)
+        partial_name, dir = popfirst!(check)
+
+        for filename in readdir(dir)
+            name = isempty(partial_name) ? filename : "$partial_name/$filename"
+            path = joinpath(dir, filename)
+
+            if isdir(path)
+                push!(check, (name, path))
+            else
+                f(name, path)
+            end
+        end
+    end
+    return nothing
+end